remove duplicate controller declaration.
This was causing the controller to execute twice.

"Note that you can also attach controllers to the DOM by declaring it
in a route definition via the $route service. A common mistake is to
declare the controller again using ng-controller in the template itself.
This will cause the controller to be attached and executed twice."

https://docs.angularjs.org/api/ng/directive/ngController